// src/configs/types/product.ts
export type TProduct = {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
};

export type TProductList = {
  products: TProduct[];
  total: number;
  skip: number;
  limit: number;
};